Added event loop for file descriptors and tasks
This patch adds an event loop that is based on file-descriptor
I/O, where users add file descriptors to the event loop. The loop
monitors file descriptors using the epoll system-call interface.
When a file descriptor becomes ready, the event loop invokes a
user-supplied call-back function for handling the event.

The event-loop's first user is the task dispatcher. A task is a
pair of function and data that runs as part of the event loop
(i.e. in the same thread). The task dispatcher creates tasks and
sends them to the event loop in a thread-safe manner, using an
I/O pipe.